Make TypeMetadataBuilder top level
